GEP 6774 API Integration for Hotlist

* Filter and sorting in get hotlist api work.
get source name work
get color code work
get audio path work
delete hotlist using predicate, avoided one extra call

* merged from development

* Solved merging issues.

* Change in sourceId and stationId fixed.


Approved-by: Mehmood Usman

+++ Source/Crossbones.ALPR.Database/Entities/SourceType.cs
@@ -11,7 +11,7 @@ public partial class SourceType
 {
     [Key]
     [Column("sysSerial")]
-    public int SysSerial { get; set; }
+    public long SysSerial { get; set; }
 
     [StringLength(50)]
     [Unicode(false)]
